RAG_ServiceProject_SingleAgent/requirements.txt




cohere==5.13.12
langchain==0.3.17
langchain-cohere==0.4.2
langchain-community==0.3.16
langchain-core==0.3.34
langchain-openai==0.3.4
langchain-pinecone==0.2.2
langchain-tests==0.3.11
langchain-text-splitters==0.3.5
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.3.6
llama-cloud==0.1.11
llama-index==0.12.16
llama-index-agent-openai==0.4.3
llama-index-cli==0.4.0
llama-index-core==0.12.16.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.17
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.20
openai==1.61.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1712154790137/work
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
PineconePDFExtractor==0.1.8
PyMuPDF==1.25.2
pymupdf4llm==0.0.17
PyMuPDFb==1.23.3
pyparsing==3.1.2
pypdf==5.2.0
tiktoken==0.8.0
uc-micro-py==1.0.3
unstructured==0.16.20
unstructured-client==0.29.0